tools/requirements.txt
typer[all]
rich
pyyaml
openai>=1.0.0
black
flake8
